requirements.txt
requests
python-firebase
geocoder
APScheduler==3.0.0
logentries